remove RandomWaveformInjection
